feat: Add Health Checks separate DI configuration and appsettings
- Introduced HealthChecksOptions and HangfireHealthCheckOptions for configurable timeouts and thresholds across all health check dependencies
- Extracted health checks setup into dedicated HealthChecksDependencyInjection module with support for database, messaging (Kafka), Redis, external APIs, and Hangfire monitoring
- Updated Kafka Avro serializer settings to use Record subject naming strategy with schema normalization disable
