+++ src/DotNetAtlas.Api/appsettings.Local.json
@@ -15,6 +15,12 @@
     }
   },
   // Infrastructure layer
+  // Communication & Messaging
+  "Kafka": {
+    "AvroSerializer": {
+      "AutoRegisterSchemas": true
+    }
+  },
   // Authentication & Authorization
   "Authentication": {
     "JwtBearer": {
