+++ src/DotNetAtlas.Api/appsettings.Testing.json
@@ -40,6 +40,11 @@
     }
   },
   // Communication & Messaging
+  "Kafka": {
+    "AvroSerializer": {
+      "AutoRegisterSchemas": true
+    }
+  },
   "KafkaForecastEventsProducer": {
     "AllowAutoCreateTopics": true
   }
