+++ src/DotNetAtlas.Api/appsettings.json
@@ -137,8 +137,9 @@
       "Url": "http://localhost:8081"
     },
     "AvroSerializer": {
-      "SubjectNameStrategy": "TopicRecord",
-      "AutoRegisterSchemas": true
+      "SubjectNameStrategy": "Record",
+      "AutoRegisterSchemas": false,
+      "NormalizeSchemas": true
     }
   },
   "Topics": {
@@ -198,6 +199,20 @@
     "FakeWeatherAlert": {
       "Cron": "*/10 * * * * *",
       "Queue": "critical"
+    }
+  },
+  "HealthChecks": {
+    "SelfTimeout": "00:00:02",
+    "SqlServerTimeout": "00:00:05",
+    "KafkaTimeout": "00:00:05",
+    "RedisTimeout": "00:00:04",
+    "IdmApiTimeout": "00:00:03",
+    "ExternalProvidersApiTimeout": "00:00:05",
+    "Hangfire": {
+      "Timeout": "00:00:03",
+      "DegradedMaximumJobsFailed": 3,
+      "UnhealthyMaximumJobsFailed": 10,
+      "MinimumAvailableServers": 1
     }
   }
 }
