+++ src/DotNetAtlas.Infrastructure/Common/Extensions/HostEnvironmentExtensions.cs
@@ -1,5 +1,3 @@
-using System.Reflection;
-using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.Hosting;
 
 namespace DotNetAtlas.Infrastructure.Common.Extensions;
