+++ src/DotNetAtlas.Infrastructure/HttpClients/WeatherProviders/OpenMeteo/OpenMeteoWeatherProvider.cs
@@ -1,11 +1,11 @@
 using System.Globalization;
 using System.Net.Http.Json;
+using DotNetAtlas.Application.WeatherForecast;
 using DotNetAtlas.Application.WeatherForecast.GetForecasts;
 using DotNetAtlas.Application.WeatherForecast.Services.Abstractions;
 using DotNetAtlas.Application.WeatherForecast.Services.Requests;
 using FluentResults;
 using Microsoft.Extensions.DependencyInjection;
-using WeatherForecastMapper = DotNetAtlas.Application.WeatherForecast.WeatherForecastMapper;
 
 namespace DotNetAtlas.Infrastructure.HttpClients.WeatherProviders.OpenMeteo;
 
@@ -30,7 +30,7 @@ public class OpenMeteoWeatherProvider : IMainWeatherForecastProvider
         ForecastRequest forecastRequest,
         CancellationToken ct)
     {
-        var geoRequest = WeatherForecastMapper.ToGeocodingRequest(forecastRequest);
+        var geoRequest = forecastRequest.ToGeocodingRequest();
         var geoResult = await _geocodingService.GetCoordinatesAsync(geoRequest, ct);
         if (geoResult.IsFailed)
         {
